TRACING-3869 Add terraform schema definition for Rum Source
